ui is working

+++ Model_Training/fetch_stock_metrics.py
@@ -21,7 +21,8 @@ def get_stock_metrics(company_df: pandas.DataFrame):
             for col in new_columns:
                 if col in ticker.info:
                     if ticker.info[col] is not None:
-                        company_df.at[index, col] = int(round(ticker.info[col], 2))
+                        # print(round(ticker.info[col], 2), type(ticker.info[col]))
+                        company_df.at[index, col] = round(ticker.info[col], 2)
                     else:
                         company_df.at[index, col] = None
